src/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
httpx
cachetools
newsapi-python
pandas
vaderSentiment
